Expose snowball-stemmers typings correctly
Needed to wrap the `@types/snowball-stemmer` typings in a module
that aligns them with the official `snowball-stemmers` package name so
that consuming applications that reference them align the types
correctly at runtime with the actual JS module.

NOTE: right now this still requires the consuming application to use
either a `typeRoots` or `path` based solution to point to
the `@types/snowball-stemmers` typings files. Still looking for full
solution to that, though the answer is probably to stop shipping the
types as embedded in this package and instead as an official,
@types standalone package.